move plot code
