+++ pydmd/dmd.py
@@ -52,7 +52,7 @@ class DMD(DMDBase):
         :param X: the input snapshots.
         :type X: numpy.ndarray or iterable
         """
-        self._snapshots, self._snapshots_shape = self._col_major_2darray(X)
+        self._snapshots = self._col_major_2darray(X)
 
         n_samples = self._snapshots.shape[1]
         X = self._snapshots[:, :-1]
