+++ pydmd/dmdoperator.py
@@ -248,24 +248,3 @@ class DMDOperator():
 
         self._modes = high_dimensional_eigenvectors
         self._Lambda = high_dimensional_eigenvalues
-
-    def plot_operator(self):
-        """
-        Plot the low-rank Atilde operator
-        """
-
-        matrix = self.as_numpy_array
-        cmatrix = matrix.real
-        rmatrix = matrix.imag
-
-        if np.linalg.norm(cmatrix) > 1.e-12:
-            _, axes = plt.subplots(nrows=1, ncols=2)
-
-            axes[0].set_title('Real')
-            axes[0].matshow(rmatrix, cmap='jet')
-            axes[1].set_title('Complex')
-            axes[1].matshow(cmatrix, cmap='jet')
-        else:
-            plt.title('Real')
-            plt.matshow(rmatrix)
-        plt.show()
