+++ tests/test_dmdc.py
@@ -1,10 +1,10 @@
 from __future__ import division
-from past.utils import old_div
-from pydmd import DMDc
-import matplotlib.pyplot as plt
+
 import numpy as np
 import scipy
 from pytest import raises
+
+from pydmd import DMDc
 
 
 def create_system_with_B():
